refactor(scintillator/nodes): extract each node to its own file
