+++ src/scintillator/nodes/lib/base.js
